creacion de materia vacia y cambio de numeros de telefono y codigos de int a long

+++ EjemploGUI/src/ejemplogui/Estudiante.java
@@ -15,7 +15,7 @@ public class Estudiante extends Persona{
     
     ArrayList<Materia> materias;
     
-    public Estudiante(String name, String lastName, int landLine, int movilPhone, String mail, String address, ArrayList<Materia> materias) {
+    public Estudiante(String name, String lastName, long landLine, long movilPhone, String mail, String address, ArrayList<Materia> materias) {
         super(name, lastName, landLine, movilPhone, mail, address);
         this.materias= materias;
     }
